# Conflicts:
#	examples/metrics/pom.xml
#	examples/pom.xml
#	examples/transform/pom.xml
#	gbif/coordinator/pom.xml
#	gbif/coordinator/tasks-integration-tests-hbase/pom.xml
#	gbif/coordinator/tasks-integration-tests/pom.xml
#	gbif/coordinator/tasks/pom.xml
#	gbif/diagnostics/pom.xml
#	gbif/keygen/pom.xml
#	gbif/pipelines/clustering-gbif/pom.xml
#	gbif/pipelines/export-gbif-hbase/pom.xml
#	gbif/pipelines/ingest-gbif-beam/pom.xml
#	gbif/pipelines/ingest-gbif-fragmenter/pom.xml
#	gbif/pipelines/ingest-gbif-java/pom.xml
#	gbif/pipelines/pom.xml
#	gbif/pipelines/pre-backbone-release/pom.xml
#	gbif/pom.xml
#	gbif/validator/pipelines-validator-ws/pom.xml
#	gbif/validator/pom.xml
#	gbif/validator/validator-api/pom.xml
#	gbif/validator/validator-checklists/pom.xml
#	gbif/validator/validator-core/pom.xml
#	gbif/validator/validator-eml/pom.xml
#	gbif/validator/validator-it-tests/pom.xml
#	gbif/validator/validator-mail/pom.xml
#	gbif/validator/validator-ws-client/pom.xml
#	livingatlas/bitmap/pom.xml
#	livingatlas/migration/pom.xml
#	livingatlas/pipelines/pom.xml
#	livingatlas/pom.xml
#	pom.xml
#	sdks/beam-common/pom.xml
#	sdks/beam-transforms/pom.xml
#	sdks/core/pom.xml
#	sdks/models/pom.xml
#	sdks/pom.xml
#	sdks/variables/pom.xml
#	tools/archives-converters/pom.xml
#	tools/elasticsearch-tools/pom.xml
#	tools/extension-converter-maven-plugin/pom.xml
#	tools/pipelines-maven-plugin/pom.xml
#	tools/pom.xml
#	tools/xml-to-avsc-maven-plugin/pom.xml

+++ gbif/coordinator/tasks/src/main/java/org/gbif/pipelines/common/GbifApi.java
@@ -21,12 +21,12 @@ public class GbifApi {
 
   /** Get number of record using Occurrence API */
   @SneakyThrows
-  public static int getIndexSize(
+  public static long getIndexSize(
       HttpClient httpClient, RegistryConfiguration config, String datasetId) {
     int nano = LocalDateTime.now().getNano();
     String url = config.wsUrl + "/occurrence/search?limit=0&datasetKey=" + datasetId + "&_" + nano;
     HttpResponse response = executeGet(httpClient, url);
-    return MAPPER.readTree(response.getEntity().getContent()).findValue("count").asInt();
+    return MAPPER.readTree(response.getEntity().getContent()).findValue("count").asLong();
   }
 
   @SneakyThrows
